engine: extract esm/layers/{reference_frame,orbit_layer}.js
